refactor content-helpers
