+++ libs/content-helpers/package.json
@@ -2,7 +2,8 @@
   "name": "@watheia/content-helpers",
   "version": "0.0.1",
   "dependencies": {
-    "@swc/helpers": "~0.5.2"
+    "@swc/helpers": "~0.5.2",
+    "@watheia/content-model": "0.0.1"
   },
   "type": "commonjs",
   "main": "./src/index.js",
