+++ libs/content-helpers/src/index.ts
@@ -1,5 +1,6 @@
 export * from './lib/content-helpers';
 export * from './lib/data-utils';
 export * from './lib/get-data-attrs';
+export * from './lib/lang-utils';
 export * from './lib/map-styles-to-class-names';
 export * from './lib/seo-utils';
